start ability to read out pixel memory
- we currently store all our data in the framebuffer,
  eventually I'll write to shadow memory as well, but for now we need to read the
  data out of VRAM

- adds call to read out pixel data from libgraphics
- does a naive RLE of it

the idea is that in the future buffers that are off screen get moved into "cold storage"
where they are compressed. we want this compression to be _fast_ not necessarily
space efficient. (presumably we are moving from limited VRAM to much more plentiful
system RAM.)

when that buffer enters the scanbox again if it's been swapped out we just uncompress it
and upload it back into VRAM.