(bump version)
this is a completely rewritten rendering subsystem.

instead of storing raster drawings we instead collect input samples and store
them twice: once in main memory and once as a collection of verts in VRAM.

this gives us a data structure that is amenable to both rasterizing and serialization.

since we aren't storing raw raster data the entire RLE/swapping system is removed.